+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/color/SingleBandOptions.scala
@@ -1,0 +1,15 @@
+package com.rasterfoundry.backsplash.color
+
+import io.circe.Json
+import io.circe.generic.JsonCodec
+
+object SingleBandOptions {
+
+  @JsonCodec
+  final case class Params(band: Int,
+                          dataType: BandDataType,
+                          colorBins: Int,
+                          colorScheme: Json,
+                          legendOrientation: String,
+                          extraNoData: Seq[Int] = Seq.empty[Int])
+}