+++ app-backend/db/src/test/scala/com/azavea/rf/database/ProjectLayerDatasourceDaoSpec.scala
@@ -55,7 +55,8 @@ class ProjectLayerDatasourceDaoSpec
                   ) => {
                 ProjectDao.addScenesToProject(
                   dbScenes map { _.id },
-                  dbProject.id
+                  dbProject.id,
+                  dbProject.defaultLayerId
                 ) flatMap { _ =>
                   {
                     ProjectLayerDatasourcesDao.listProjectLayerDatasources(